Move MapEntity to d2interface
